Extract BaseTransactionsService
